fix: delete resources if realm is already gone

+++ internal/controller/group_controller.go
@@ -59,6 +59,10 @@ func (r *GroupReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl
 
 	session, err := KeycloakSessionForRealm(ctx, r.Client, instance.Spec.RealmRef, req.Namespace)
 	if err != nil {
+		if !instance.DeletionTimestamp.IsZero() && apierrors.IsNotFound(err) {
+			log.Info("Realm no longer exists, skipping Keycloak cleanup", "error", err.Error())
+			return r.handleDeletion(ctx, instance, nil)
+		}
 		log.Error(err, "Unable to get Keycloak session")
 		return ctrl.Result{RequeueAfter: FailedKeycloakConnectionRetryPeriod}, nil
 	}
@@ -85,7 +89,7 @@ func (r *GroupReconciler) handleDeletion(ctx context.Context, instance *v1alpha1
 		return ctrl.Result{}, nil
 	}
 
-	if instance.Status.KeycloakID != nil {
+	if session != nil && instance.Status.KeycloakID != nil {
 		log.Info("Deleting group from Keycloak", "id", *instance.Status.KeycloakID)
 		if err := session.Client.DeleteGroup(ctx, session.Token.AccessToken, session.RealmName, *instance.Status.KeycloakID); err != nil {
 			return ctrl.Result{}, fmt.Errorf("failed to delete group from Keycloak: %w", err)
