+++ internal/controller/openid/openidclient_controller.go
@@ -61,6 +61,10 @@ func (r *OpenIDClientReconciler) Reconcile(ctx context.Context, req ctrl.Request
 
 	session, err := controller.KeycloakSessionForRealm(ctx, r.Client, instance.Spec.RealmRef, req.Namespace)
 	if err != nil {
+		if !instance.DeletionTimestamp.IsZero() && apierrors.IsNotFound(err) {
+			log.Info("Realm no longer exists, skipping Keycloak cleanup", "error", err.Error())
+			return r.handleDeletion(ctx, instance, nil)
+		}
 		log.Error(err, "Unable to get Keycloak session")
 		return ctrl.Result{RequeueAfter: controller.FailedKeycloakConnectionRetryPeriod}, nil
 	}
@@ -87,7 +91,7 @@ func (r *OpenIDClientReconciler) handleDeletion(ctx context.Context, instance *v
 		return ctrl.Result{}, nil
 	}
 
-	if instance.Status.KeycloakID != nil {
+	if session != nil && instance.Status.KeycloakID != nil {
 		log.Info("Deleting client from Keycloak", "id", *instance.Status.KeycloakID)
 		if err := session.Client.DeleteClient(ctx, session.Token.AccessToken, session.RealmName, *instance.Status.KeycloakID); err != nil {
 			var apiErr *gocloak.APIError
